:hammer: Move Text types into Text namespace
