+++ yarhl.UnitTests/Media/Text/Po2BinaryTests.cs
@@ -23,13 +23,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.Media
+namespace Yarhl.UnitTests.Media.Text
 {
     using System;
     using NUnit.Framework;
     using Yarhl.FileFormat;
     using Yarhl.IO;
-    using Yarhl.Media;
+    using Yarhl.Media.Text;
 
     [TestFixture]
     public class Po2BinaryTests
