+++ yarhl.UnitTests/Media/Text/PoHeaderTests.cs
@@ -23,11 +23,11 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.Media
+namespace Yarhl.UnitTests.Media.Text
 {
     using System;
     using NUnit.Framework;
-    using Yarhl.Media;
+    using Yarhl.Media.Text;
 
     [TestFixture]
     public class PoHeaderTests
