+++ yarhl.UnitTests/Media/Text/XmlExtensionTests.cs
@@ -23,12 +23,12 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.Media
+namespace Yarhl.UnitTests.Media.Text
 {
     using System;
     using System.Xml.Linq;
     using NUnit.Framework;
-    using Yarhl.Media;
+    using Yarhl.Media.Text;
 
     [TestFixture]
     public class XmlExtensionTests
